+++ packages/api/src/__generated__/schema.ts
@@ -129,6 +129,16 @@ export type AvailableDeliveryWindows = {
   tax?: Maybe<Scalars['Int']>;
 };
 
+export type BusinessHour = {
+  __typename?: 'BusinessHour';
+  /** Business hour closing time. */
+  ClosingTime?: Maybe<Scalars['String']>;
+  /** Number that represents the day of the week. */
+  DayOfWeek?: Maybe<Scalars['Int']>;
+  /** Business hour opening time. */
+  OpeningTime?: Maybe<Scalars['String']>;
+};
+
 export type DeliveryIds = {
   __typename?: 'DeliveryIds';
   /** DeliveryIds courier id */
@@ -169,10 +179,8 @@ export type IShippingItem = {
 };
 
 export type IStoreB2B = {
-  corporateName?: Maybe<Scalars['String']>;
   customerId: Scalars['String'];
   firstName?: Maybe<Scalars['String']>;
-  isCorporate?: Maybe<Scalars['Boolean']>;
   isRepresentative?: Maybe<Scalars['Boolean']>;
   lastName?: Maybe<Scalars['String']>;
   unitId?: Maybe<Scalars['String']>;
@@ -346,6 +354,14 @@ export type IUserOrderCancel = {
   reason?: Maybe<Scalars['String']>;
 };
 
+export type Item = {
+  __typename?: 'Item';
+  /** Pickup point distance. */
+  distance?: Maybe<Scalars['Float']>;
+  /** Pickup point. */
+  pickupPoint?: Maybe<PickupPoint>;
+};
+
 export type LogisticsInfo = {
   __typename?: 'LogisticsInfo';
   /** LogisticsInfo itemIndex. */
@@ -444,6 +460,18 @@ export type MutationValidateSessionArgs = {
   session: IStoreSession;
 };
 
+export type Paging = {
+  __typename?: 'Paging';
+  /** Current page. */
+  page?: Maybe<Scalars['Int']>;
+  /** Number of items per page. */
+  pageSize?: Maybe<Scalars['Int']>;
+  /** Total number of pages. */
+  pages?: Maybe<Scalars['Int']>;
+  /** Total number of items. */
+  total?: Maybe<Scalars['Int']>;
+};
+
 /** Newsletter information. */
 export type PersonNewsletter = {
   __typename?: 'PersonNewsletter';
@@ -479,6 +507,60 @@ export type PickupAddress = {
   state?: Maybe<Scalars['String']>;
   /** PickupAddress street. */
   street?: Maybe<Scalars['String']>;
+};
+
+export type PickupPoint = {
+  __typename?: 'PickupPoint';
+  /** Pickup point additional info. */
+  additionalInfo?: Maybe<Scalars['String']>;
+  /** Pickup point address. */
+  address?: Maybe<PickupPointAddress>;
+  /** Pickup point business hours. */
+  businessHours?: Maybe<BusinessHour>;
+  /** Pickup point friendly name. */
+  friendlyName?: Maybe<Scalars['String']>;
+  /** Pickup point id. */
+  id?: Maybe<Scalars['String']>;
+};
+
+export type PickupPointAddress = {
+  __typename?: 'PickupPointAddress';
+  /** Address id. */
+  addressId?: Maybe<Scalars['String']>;
+  /** Address type. */
+  addressType?: Maybe<Scalars['String']>;
+  /** Address city. */
+  city?: Maybe<Scalars['String']>;
+  /** Address complement */
+  complement?: Maybe<Scalars['String']>;
+  /** Address country. */
+  country?: Maybe<Scalars['String']>;
+  /** Address geo coordinates. */
+  geoCoordinates?: Maybe<Array<Maybe<Scalars['Float']>>>;
+  /** Inform whether the address is disposable. */
+  isDisposable?: Maybe<Scalars['Boolean']>;
+  /** Address neighborhood. */
+  neighborhood?: Maybe<Scalars['String']>;
+  /** Address number. */
+  number?: Maybe<Scalars['String']>;
+  /** Address postal code. */
+  postalCode?: Maybe<Scalars['String']>;
+  /** Address receiver name. */
+  receiverName?: Maybe<Scalars['String']>;
+  /** Address reference. */
+  reference?: Maybe<Scalars['String']>;
+  /** Address state. */
+  state?: Maybe<Scalars['String']>;
+  /** Address street. */
+  street?: Maybe<Scalars['String']>;
+};
+
+export type PickupPoints = {
+  __typename?: 'PickupPoints';
+  /** List of pickup points of the given location. */
+  items?: Maybe<Array<Maybe<Item>>>;
+  /** Pagination details. */
+  paging?: Maybe<Paging>;
 };
 
 export type PickupStoreInfo = {
@@ -547,6 +629,8 @@ export type Query = {
   collection: StoreCollection;
   /** Returns information about the list of Orders that the User can view. */
   listUserOrders?: Maybe<UserOrderListMinimalResult>;
+  /** Returns a list of pickup points near to the given geo coordinates or postal code + country code. */
+  pickupPoints?: Maybe<PickupPoints>;
   /** Returns the details of a product based on the specified locator. */
   product: StoreProduct;
   /** Returns the total product count information based on a specific location accessible through the VTEX segment cookie. */
@@ -591,6 +675,13 @@ export type QueryListUserOrdersArgs = {
   perPage?: Maybe<Scalars['Int']>;
   status?: Maybe<Array<Maybe<Scalars['String']>>>;
   text?: Maybe<Scalars['String']>;
+};
+
+
+export type QueryPickupPointsArgs = {
+  country?: Maybe<Scalars['String']>;
+  geoCoordinates?: Maybe<IStoreGeoCoordinates>;
+  postalCode?: Maybe<Scalars['String']>;
 };
 
 
@@ -795,10 +886,8 @@ export type StoreAuthor = {
 
 export type StoreB2B = {
   __typename?: 'StoreB2B';
-  corporateName?: Maybe<Scalars['String']>;
   customerId: Scalars['String'];
   firstName?: Maybe<Scalars['String']>;
-  isCorporate?: Maybe<Scalars['Boolean']>;
   isRepresentative?: Maybe<Scalars['Boolean']>;
   lastName?: Maybe<Scalars['String']>;
   unitId?: Maybe<Scalars['String']>;
@@ -1662,6 +1751,7 @@ export type UserOrderFromList = {
   clientName?: Maybe<Scalars['String']>;
   creationDate?: Maybe<Scalars['String']>;
   currencyCode?: Maybe<Scalars['String']>;
+  customFields?: Maybe<UserOrderFromListCustomFields>;
   deliveryDates?: Maybe<Array<Maybe<Scalars['String']>>>;
   giftCardProviders?: Maybe<Array<Maybe<Scalars['String']>>>;
   hostname?: Maybe<Scalars['String']>;