+++ packages/api/src/platforms/vtex/resolvers/query.ts
@@ -11,6 +11,7 @@ import type {
   QuerySellersArgs,
   QueryShippingArgs,
   QueryUserOrderArgs,
+  QueryPickupPointsArgs,
 } from '../../../__generated__/schema'
 import { BadRequestError, ForbiddenError, NotFoundError } from '../../errors'
 import type { CategoryTree } from '../clients/commerce/types/CategoryTree'
@@ -451,10 +452,27 @@ export const Query = {
         status: order.status,
         statusDescription: order.statusDescription,
         ShippingEstimatedDate: order.ShippingEstimatedDate,
-        // customFields: order.customFields,
+        customFields: order.customFields,
         currencyCode: order.currencyCode,
       })),
       paging: orders.paging,
     }
+  },
+  pickupPoints: async (
+    _: unknown,
+    { country, postalCode, geoCoordinates }: QueryPickupPointsArgs,
+    ctx: Context
+  ) => {
+    const {
+      clients: { commerce },
+    } = ctx
+
+    const result = await commerce.checkout.pickupPoints({
+      country,
+      postalCode,
+      geoCoordinates,
+    })
+
+    return result
   },
 }