+++ packages/core/src/pages/account/orders/index.tsx
@@ -95,6 +95,12 @@ const query = gql(`
         statusDescription
         ShippingEstimatedDate
         currencyCode
+        customFields {
+          costCenter
+          poNumber
+          release
+          desktop
+        }
       }
       paging {
         total