+++ packages/core/test/server/index.test.ts
@@ -76,6 +76,7 @@ const QUERIES = [
   'productCount',
   'userOrder',
   'listUserOrders',
+  'pickupPoints',
 ]
 
 const MUTATIONS = [